* Used suggestion by G.W. Reynolds to raise murphymax values for pol5 to   avoid memory issues in pol51opt.  It should have the added bonus of   getting to the good polynomials faster.  However, there are still many   instances remaining where pol51m0b dies with the message "lambda-comp"   or "ull5a5p-comp". * Moved -wt option help text from matsolve to matbuild.  Also updated the   Perl script to use the -wt option with matbuild, not matsolve.  Allowed   "matsolve --help" to print help text if a depinf file does not exist. * Added support in the Perl script of lss and mrif parameters in the .poly   file.  They correspond to $LATSIEVE_SIDE and $maxRelsInFF settings. * Tweaked some Makefiles so I don't always have to edit them extensively   to use them in my environment. :) * Fixed FLOAT64 define in pol51m0b.c * Suppressed warning in montgomery_sqrt.c with a couple typecasts to s32s.   There really needs to be a more portable way to printf a 64-bit int.
git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@17 80ef69ca-9341-0410-9d60-8acc7774fa31

+++ src/montgomery_sqrt.c
@@ -985,7 +985,7 @@ ABexponentSum=0;
       s32 pFacts[128];
       int numpFacts;
       printf("Factorization of relation %ld is wrong:\n", i);
-      printf("a=%ld, b=%ld\n", a, -b);
+      printf("a=%ld, b=%ld\n", (s32)a, (s32)-b);
       printf("a-bm = "); mpz_out_str(stdout, 10, tmp); printf("\n");
       printf("Stored large primes are: %ld %ld.\n", (u32)R.p[0], (u32)R.p[1]);
       printf("Product of factors gives:\n       ");
